+++ changelog.md
@@ -1,6 +1,7 @@
 #### 0.0.5 (2021-09-29)
 
-- Set default values for `acfml_lang_active` (#e179730)
+- add 'acfml_ui_style' = 'simple' (#744c894)
+- Set default values for `acfml_lang_active` (#d739ebe)
 - Performance Optimizations (#77d064e)
 - fix for `remove_default_language_from_url` (#1262fb8)
 - add clearer redirect logic (#a75f634)